% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_effects.R
\name{adj_effects}
\alias{adj_effects}
\title{Calculate variant-phenotype associations adjusted for another phenotype using summary data}
\usage{
adj_effects(
  beta1,
  se1,
  beta2,
  se2,
  eaf1,
  eaf2,
  n1,
  n2,
  corr,
  varY1 = NULL,
  varY2 = NULL
)
}
\arguments{
\item{varY2}{}
}
\value{

}
\description{
Calculate variant-phenotype associations adjusted for another phenotype using summary data
}
