% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_summarystats.R
\name{adjusted_summarystats}
\alias{adjusted_summarystats}
\title{Generate variant-phenotype summary statistics adjusted for another phenotype using summary data}
\usage{
adjusted_summarystats(
  beta1,
  se1,
  eaf1,
  n1,
  var1,
  beta2,
  se2,
  eaf2,
  n2,
  var2,
  corr = NULL
)
}
\arguments{
\item{corr}{}
}
\value{

}
\description{
Generate variant-phenotype summary statistics adjusted for another phenotype using summary data
}
