% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_results}
\alias{check_results}
\title{Helper function to check multivariable colocalization results}
\usage{
check_results(res, type = "cont", thresh1 = 0.5, thresh2 = 0.5)
}
\arguments{
\item{thresh2}{}
}
\value{

}
\description{
Helper function to check multivariable colocalization results
}
