% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sumstats.R
\name{generate_sumstats}
\alias{generate_sumstats}
\title{Generate summary statistics for a continuous phenotype}
\usage{
generate_sumstats(phenotype, x, adj = NULL)
}
\arguments{
\item{adj}{}
}
\value{

}
\description{
Generate summary statistics for a continuous phenotype
}
