% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sumstats_logistic.R
\name{generate_sumstats_logistic}
\alias{generate_sumstats_logistic}
\title{Generate summary statistics for a binary phenotype}
\usage{
generate_sumstats_logistic(phenotype, x, adj = NULL)
}
\arguments{
\item{adj}{}
}
\value{

}
\description{
Generate summary statistics for a binary phenotype
}
