% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotypes.R
\name{simulate_phenotypes}
\alias{simulate_phenotypes}
\title{Simulate a phenotype}
\usage{
simulate_phenotypes(
  seed,
  genotype_data1,
  genotype_data2,
  r2_lower,
  r2_upper,
  h.exp1,
  h.exp2,
  theta1,
  theta2,
  h.out,
  theta.vec,
  ybin_prev = 0.5
)
}
\arguments{
\item{ybin_prev}{}
}
\value{

}
\description{
Simulate a phenotype
}
