% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotypes_null.R
\name{simulate_phenotypes_null}
\alias{simulate_phenotypes_null}
\title{Simulate phenotypes, null associations (NOTE: not needed)}
\usage{
simulate_phenotypes_null(
  seed,
  genotype_data1,
  genotype_data2,
  r2_lower,
  r2_upper,
  h.exp1,
  h.exp2,
  theta1,
  theta2,
  h.out,
  theta.vec
)
}
\arguments{
\item{theta.vec}{}
}
\value{

}
\description{
Simulate phenotypes, null associations (NOTE: not needed)
}
